Tolerate secrets being missing
If any secret was missing, the reconcile would error out. After doing
this a number of times, the controller would stop trying and end up in a
state where the operator would need restarting, or the resource would
need recreating, to continue reconciliation.

This fix stops the reconcile from erroring out so the operator will
reconcile continuously until the secrets are all present.

Fixes #22170